server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
websockets==13.0
openai==1.46.0
httpx==0.27.2
